Added the MinMax Algorithm with Alpha Beta Pruning for an intelligent chess bot.

+++ src/Headers/Bot.h
@@ -30,7 +30,7 @@ public:
     Move getIntermediateMove();
     Move getMinMaxMove(int depth);
     Move getHardMove();
-    int minMaxAlgo(Board board, Move & move, int depth, int alpha, int beta, bool maximizing);
+    float minMaxAlgo(Board board, Move & move, int depth, float alpha, float beta, bool maximizing);
 };
